# optimize repeated container creation in tests