--- a/src/Testing/TestCase.neon
+++ b/src/Testing/TestCase.neon
@@ -9,6 +9,9 @@ services:
 		arguments:
 			phpParser: @phpParserDecorator
 			php8Parser: @php8PhpParser
+			fileExtensions: %fileExtensions%
+			obsoleteExcludesAnalyse: %excludes_analyse%
+			excludePaths: %excludePaths%
 
 	cacheStorage:
 		class: PHPStan\Cache\MemoryCacheStorage