--- a/src/Testing/TestCaseSourceLocatorFactory.php
+++ b/src/Testing/TestCaseSourceLocatorFactory.php
@@ -12,6 +12,7 @@ use PHPStan\BetterReflection\SourceLocator\Type\EvaledCodeSourceLocator;
 use PHPStan\BetterReflection\SourceLocator\Type\MemoizingSourceLocator;
 use PHPStan\BetterReflection\SourceLocator\Type\PhpInternalSourceLocator;
 use PHPStan\BetterReflection\SourceLocator\Type\SourceLocator;
+use PHPStan\Php\PhpVersion;
 use PHPStan\Reflection\BetterReflection\SourceLocator\AutoloadSourceLocator;
 use PHPStan\Reflection\BetterReflection\SourceLocator\ComposerJsonAndInstalledJsonSourceLocatorMaker;
 use PHPStan\Reflection\BetterReflection\SourceLocator\FileNodesFetcher;
@@ -19,10 +20,20 @@ use PHPStan\Reflection\BetterReflection\SourceLocator\PhpVersionBlacklistSourceL
 use ReflectionClass;
 use function dirname;
 use function is_file;
+use function serialize;
+use function sha1;
 
 class TestCaseSourceLocatorFactory
 {
 
+	/** @var array<string, list<SourceLocator>> */
+	private static array $composerSourceLocatorsCache = [];
+
+	/**
+	 * @param string[] $fileExtensions
+	 * @param string[] $obsoleteExcludesAnalyse
+	 * @param array{analyse?: array<int, string>, analyseAndScan?: array<int, string>}|null $excludePaths
+	 */
 	public function __construct(
 		private ComposerJsonAndInstalledJsonSourceLocatorMaker $composerJsonAndInstalledJsonSourceLocatorMaker,
 		private Parser $phpParser,
@@ -30,6 +41,10 @@ class TestCaseSourceLocatorFactory
 		private FileNodesFetcher $fileNodesFetcher,
 		private PhpStormStubsSourceStubber $phpstormStubsSourceStubber,
 		private ReflectionSourceStubber $reflectionSourceStubber,
+		private PhpVersion $phpVersion,
+		private array $fileExtensions,
+		private array $obsoleteExcludesAnalyse,
+		private ?array $excludePaths,
 	)
 	{
 	}
@@ -38,8 +53,14 @@ class TestCaseSourceLocatorFactory
 	{
 		$classLoaders = ClassLoader::getRegisteredLoaders();
 		$classLoaderReflection = new ReflectionClass(ClassLoader::class);
-		$locators = [];
-		if ($classLoaderReflection->hasProperty('vendorDir')) {
+		$cacheKey = sha1(serialize([
+			$this->phpVersion->getVersionId(),
+			$this->fileExtensions,
+			$this->obsoleteExcludesAnalyse,
+			$this->excludePaths,
+		]));
+		if ($classLoaderReflection->hasProperty('vendorDir') && ! isset(self::$composerSourceLocatorsCache[$cacheKey])) {
+			$composerLocators = [];
 			$vendorDirProperty = $classLoaderReflection->getProperty('vendorDir');
 			$vendorDirProperty->setAccessible(true);
 			foreach ($classLoaders as $classLoader) {
@@ -52,10 +73,13 @@ class TestCaseSourceLocatorFactory
 				if ($composerSourceLocator === null) {
 					continue;
 				}
-				$locators[] = $composerSourceLocator;
+				$composerLocators[] = $composerSourceLocator;
 			}
+
+			self::$composerSourceLocatorsCache[$cacheKey] = $composerLocators;
 		}
 
+		$locators = self::$composerSourceLocatorsCache[$cacheKey] ?? [];
 		$astLocator = new Locator($this->phpParser);
 		$astPhp8Locator = new Locator($this->php8Parser);